backend/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
awsebcli==3.19.3
bcrypt==3.2.0
blessed==1.17.12
boto3==1.16.0
botocore==1.19.0
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.5
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-polymorphic==3.0.0
django-rest-knox==4.1.0
django-rest-polymorphic==0.1.9
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-serializer-extensions==2.0.1
docker==4.4.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
hashids==1.3.1
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
jmespath==0.10.0
jsonschema==3.2.0
msgpack==1.0.2
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.4
semantic-version==2.5.0
sentry-sdk==0.20.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.4.0
zope.interface==5.2.0
django-import-export==2.5.0